+++ flutter-ecom/lib/features/models/response/get_carts.dart
@@ -27,9 +27,9 @@ class GetCartsResponse {
 class Cart {
   String? id;
   String? productId;
-  int? price;
+  double? price;
   int? quantity;
-  int? totalPrice;
+  double? totalPrice;
   String? createdAt;
   String? updatedAt;
 
@@ -45,9 +45,9 @@ class Cart {
   Cart.fromJson(Map<String, dynamic> json) {
     id = json['id'];
     productId = json['product_id'];
-    price = json['price'];
+    price = json['price'].toDouble();
     quantity = json['quantity'];
-    totalPrice = json['total_price'];
+    totalPrice = json['total_price'].toDouble();
     createdAt = json['created_at'];
     updatedAt = json['updated_at'];
   }
